+++ Assets/MirrorMarketSDK/Runtime/Models/Request/Approve/ApproveTransferSPLToken.cs
@@ -8,7 +8,7 @@ namespace MirrorworldSDK.Wrapper
     public class ApproveTransferSPLToken
     {
         public string to_publickey;
-        public ulong amount;
+        public double amount;
         public string token_mint;
         public int decimals;
     }
