+++ Assets/MirrorMarketSDK/Runtime/Services/Interfaces/IWalletService.cs
@@ -12,7 +12,7 @@ namespace MirrorworldSDK.Interfaces
 
         public void GetWalletTransactionsBySignatrue(string signature, Action<CommonResponse<TransferTokenResponse>> action);
 
-        public void TransferSol(ulong amout, string publicKey,string confirmation, Action<CommonResponse<TransferSolResponse>> callBack);
+        public void TransferSol(int amout, string publicKey,string confirmation, Action<CommonResponse<TransferSolResponse>> callBack);
         
         public void TransferSPLToken(string token_mint, int decimals, ulong amount, string to_publickey, Action<CommonResponse<TransferTokenResponse>> callBack);
